Move path audio logic to path actions
